Move validate_backtest_data and get_timeframe to histoyr
